requirements.txt
# Python 3.8+
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-core==2.16.3
typing-extensions>=4.9.0
supabase==1.0.3
httpx>=0.23.0,<0.24.0
python-dotenv==1.0.1
openai==1.14.2